jordis merged into one
