+++ jordi/.dockerignore
