+++ jordi/.env
@@ -1,0 +1,3 @@
+MONGODB_URI="mongodb://localhost:27017/questions"
+GENERATE_ON_STARTUP="false"
+SCHEDULE="* * * 1 * *"
