--- jordi/jordi-ask/.env
@@ -1,1 +1,0 @@
-MONGODB_URI="mongodb://localhost:27017/questions"
