--- jordi/jordi-think/.dockerignore
@@ -1,3 +1,0 @@
-node_modules
-.env
-coverage
