--- jordi/jordi-think/Dockerfile
@@ -1,20 +1,0 @@
-# Use an official Node.js runtime as a parent image
-FROM node:20
-
-# Set the working directory in the container
-WORKDIR /usr/src/jordi-think
-
-# Copy package.json and package-lock.json to the working directory
-COPY package*.json ./
-
-# Install app dependencies
-RUN npm install
-
-# Copy the app source code to the working directory
-COPY . .
-
-# Expose the port the app runs on
-EXPOSE 8003
-
-# Define the command to run your app
-CMD ["node", "jordi-think-service.js"]
