--- jordi/jordi-think/package.json
@@ -1,31 +1,0 @@
-{
-    "name": "jordithinkservice",
-    "version": "1.0.0",
-    "description": " Question generator service, in charge of genearting wikidata-based Q&A chunks",
-    "author": "wiq_es05b",
-    "private": true,
-    "main": "jordi-think-service.js",
-    "scripts": {
-        "start": "node jordi-think-service.js",
-        "test": "jest"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/arquisoft/wiq_0.git"
-    },
-    "bugs": {
-        "url": "https://github.com/arquisoft/wiq_0/issues"
-    },
-    "homepage": "https://github.com/arquisoft/wiq_0#readme",
-    "dependencies": {
-        "axios": "^1.6.7",
-        "express": "^4.18.3",
-        "mongoose": "^8.0.4",
-        "node-cron": "^3.0.3"
-    },
-    "devDependencies": {
-        "jest": "^29.7.0",
-        "mongodb-memory-server": "^9.1.5",
-        "supertest": "^6.3.4"
-    }
-}
